+++ libp2p/security/insecure/transport.py
@@ -1,19 +1,22 @@
-import asyncio
 from typing import cast
 
 from libp2p.network.connection.raw_connection_interface import IRawConnection
 from libp2p.peer.id import ID
 from libp2p.security.secure_conn_interface import ISecureConn
 from libp2p.security.secure_transport_interface import ISecureTransport
+from libp2p.security.typing import TSecurityDetails
 
-from .typing import TSecurityDetails
 
+class InsecureTransport(ISecureTransport):
+    """
+    ``InsecureTransport`` provides the "identity" upgrader for a ``IRawConnection``,
+    i.e. the upgraded transport does not add any additional security.
+    """
 
-class SimpleSecurityTransport(ISecureTransport):
-    key_phrase: str
+    transport_id: str
 
-    def __init__(self, key_phrase: str) -> None:
-        self.key_phrase = key_phrase
+    def __init__(self, transport_id: str) -> None:
+        self.transport_id = transport_id
 
     async def secure_inbound(self, conn: IRawConnection) -> ISecureConn:
         """
@@ -21,16 +24,8 @@ class SimpleSecurityTransport(ISecureTransport):
         for an inbound connection (i.e. we are not the initiator)
         :return: secure connection object (that implements secure_conn_interface)
         """
-        await conn.write(self.key_phrase.encode())
-        incoming = (await conn.read()).decode()
-
-        if incoming != self.key_phrase:
-            raise Exception(
-                "Key phrase differed between nodes. Expected " + self.key_phrase
-            )
-
-        secure_conn = SimpleSecureConn(conn, self.key_phrase)
-        return secure_conn
+        insecure_conn = InsecureConn(conn, self.transport_id)
+        return insecure_conn
 
     async def secure_outbound(self, conn: IRawConnection, peer_id: ID) -> ISecureConn:
         """
@@ -38,31 +33,18 @@ class SimpleSecurityTransport(ISecureTransport):
         for an inbound connection (i.e. we are the initiator)
         :return: secure connection object (that implements secure_conn_interface)
         """
-        await conn.write(self.key_phrase.encode())
-        incoming = (await conn.read()).decode()
-
-        # Force context switch, as this security transport is built for testing locally
-        # in a single event loop
-        await asyncio.sleep(0)
-
-        if incoming != self.key_phrase:
-            raise Exception(
-                "Key phrase differed between nodes. Expected " + self.key_phrase
-            )
-
-        secure_conn = SimpleSecureConn(conn, self.key_phrase)
-        return secure_conn
+        insecure_conn = InsecureConn(conn, self.transport_id)
+        return insecure_conn
 
 
-class SimpleSecureConn(ISecureConn):
+class InsecureConn(ISecureConn):
     conn: IRawConnection
-    key_phrase: str
     details: TSecurityDetails
 
-    def __init__(self, conn: IRawConnection, key_phrase: str) -> None:
+    def __init__(self, conn: IRawConnection, conn_id: str) -> None:
         self.conn = conn
         self.details = cast(TSecurityDetails, {})
-        self.details["key_phrase"] = key_phrase
+        self.details["id"] = conn_id
 
     def get_conn(self) -> IRawConnection:
         """
