+++ tests/security/test_security_multistream.py
@@ -1,13 +1,12 @@
 import asyncio
 
-import multiaddr
-import pytest
-
 from libp2p import new_node
 from libp2p.peer.peerinfo import info_from_p2p_addr
 from libp2p.protocol_muxer.multiselect_client import MultiselectClientError
-from libp2p.security.insecure_security import InsecureTransport
-from libp2p.security.simple_security import SimpleSecurityTransport
+from libp2p.security.insecure.transport import InsecureTransport
+from libp2p.security.simple.transport import SimpleSecurityTransport
+import multiaddr
+import pytest
 from tests.utils import cleanup, connect
 
 # TODO: Add tests for multiple streams being opened on different
